[#289] Refactor: Move Method; Move logic to view model instead
